refactor(hr): extract the VS-side batched update handling into VisualStudioFileUpdater (spec 052)
EntryPoint keeps only the message routing; the batched apply / workspace
readiness / deferred saves / EnC trigger logic — and the apply-file-content
helper shared with the legacy per-file path — move to VisualStudioFileUpdater,
the IDE-side counterpart of the dev-server's IdeFileUpdater.
